feat: add stacked cost-over-time chart to the costs explorer
Extract the billing token-usage panel into a shared StackedTimeSeriesPanel
and reuse it on the costs page for week-over-week spend comparison (DNO-279).
