+++ client/dashboard/src/components/billing/tum-usage-card.tsx
@@ -2,8 +2,8 @@ import { useState } from "react";
 import { Info } from "lucide-react";
 import { SimpleTooltip } from "@/components/ui/tooltip";
 import { cn } from "@/lib/utils";
+import { ToggleButton } from "@/components/ui/toggle-button";
 import { type BillingCycle } from "./billing-cycles";
-import { ToggleButton } from "./toggle-button";
 
 const HOUR_MS = 60 * 60 * 1000;
